# openapi.yml
openapi: 3.1.0
info:
  title: Parkovi hrvatske
  description: API za parkove hrvatske
  contact:
    name: Karla Šmuk
  license:
    name: CC0-1.0 license
    url: https://creativecommons.org/publicdomain/zero/1.0/
  version: 1.0.0
paths:
  /api/parks:
    get:
      summary: Fetch all parks
      description: Fetch all parks
      operationId: listAllParks
      responses:
        '200':
          description: Fetched all parks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
    post:
      summary: Add a new park
      description: Add a new park
      operationId: addNewPark
      requestBody:
        description: Create a new park
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParkRequest'
        required: true
      responses:
        '200':
          description: New park created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'    
        '409':
          description: Park already exists 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
  /api/parks/{parkId}:
    get:
      summary: Find park by id
      description: Returns a park
      operationId: getParkByID
      parameters:
        - name: parkId
          in: path
          description: ID of park to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Fetched park with parkId
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Response'          
        '404':
          description: Park not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'  
    put:
      summary: Update an existing park
      description: Update an existing park by Id
      operationId: updatePark
      parameters:
        - name: parkId
          in: path
          description: ID of park to return
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existing park
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParkRequest'
        required: true
      responses:
        '200':
          description: Park updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '404':
          description: Park not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'  
    delete:
      summary: Delete a park
      description: Delete a park
      operationId: deletePark
      parameters:
        - name: parkId
          in: path
          description: ID of park to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Park deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '404':
          description: Park doesnt exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'  
  /api/parks/{parkId}/animals:
    get:
      summary: Find animals of park
      description: Returns animals of park
      operationId: getAnimalsByPark
      parameters:
        - name: parkId
          in: path
          description: ID of park
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Fetched animals for park with parkId
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Response'          
        '404':
          description: Park not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'  
  /api/parks/{parkId}/counties:
    get:
      summary: Find counties of park
      description: Returns counties of park
      operationId: getCountiesByPark
      parameters:
        - name: parkId
          in: path
          description: ID of park
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Fetched counties for park with parkId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'          
        '404':
          description: Park not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'  
  /api/parks/{parkId}/peak:
    get:
      summary: Find peak of park
      description: Returns peak of park
      operationId: getPeakByParkID
      parameters:
        - name: parkId
          in: path
          description: ID of park
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Fetched peak for park with parkId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '404':
          description: Peak not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'  
  /api:
    get:
      summary: API
      description: Returns openapi specification
      operationId: getAPI
      responses:
        '200':
          description: Successful response
components:
  schemas:
    Response:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        response:
          type: object
          items:
            $ref: '#/components/schemas/ParkResponse'
    ResponseError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        response:
          type: string
    CreateParkRequest:
      type: object
      properties:
        parkName:
          type: string
        typeOfParkName:
          type: string
        yearOfFoundation:
          type: integer
        area:
          type: number
          format: double
        peak:
          type: object
          items:
            $ref: '#/components/schemas/HighestPeakDto'
        counties:
          type: array
          format: list
          items:
            $ref: '#/components/schemas/CountyDto'
        atraction:
          type: string
        event:
          type: string
        animals:
          type: array
          format: list
          items:
            $ref: '#/components/schemas/AnimalDto'
    ParkResponse:
      type: object
      properties:
        parkId:
          type: integer
        parkName:
          type: string
        typeOfParkName:
          type: string
        yearOfFoundation:
          type: integer
        area:
          type: number
          format: double
        peak:
          type: object
          items:
            $ref: '#/components/schemas/HighestPeakResponse'
        counties:
          type: array
          format: list
          items:
            $ref: '#/components/schemas/CountyResponse'
        atraction:
          type: string
        event:
          type: string
        animals:
          type: array
          format: list
          items:
            $ref: '#/components/schemas/AnimalResponse'
    HighestPeakResponse:
      type: object
      properties:
        peakID:
          type: integer
        peakName:
          type: string
        peakHeight:
          type: integer
    AnimalResponse:
      type: object
      properties:
        animalID:
          type: integer
        animalName:
          type: string
        speciesOfAnimal:
          type: string
    CountyResponse:
      type: object
      properties:
        countyID:
          type: integer
        countyName:
          type: string
    HighestPeakDto:
      type: object
      properties:
        peakName:
          type: string
        peakHeight:
          type: integer
    AnimalDto:
      type: object
      properties:
        animalName:
          type: string
        speciesOfAnimal:
          type: string
    CountyDto:
      type: object
      properties:
        countyName:
          type: string
  requestBodies:
    Park:
      description: Park object that needs to be created or updated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateParkRequest'